requirements.txt
flask~=2.2.3
structlog~=22.3.0
pygtail~=0.14.0
boto3~=1.26.84
minio~=7.1.13
python-dotenv~=1.0.0
smart_open
pandas~=1.5.3
numpy~=1.23.1
spacy~=3.5.0
unidecode~=1.3.6
contractions~=0.1.73
word2number~=1.1
gensim~=4.3.0
beautifulsoup4~=4.11.2
nltk~=3.8.1
transformers~=4.26.1
torch~=1.13.1
torchaudio~=0.13.1
soundfile~=0.12.1
botocore~=1.29.84
requests~=2.28.2
sentencepiece~=0.1.97
protobuf~=3.20.0
bcrypt~=4.0.1
opencv-python~=4.6.0.66
pynput~=1.7.6
tqdm~=4.64.1
bs4~=0.0.1